server/requirements.txt
# Backend для детектора дефектов с Anomalib (Intel)
# Python 3.10+

anomalib>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pillow>=10.0.0
requests>=2.28.0
# pandas 3.x ломает логирование в Lightning/Anomalib ("cannot set a frame with no defined index and a scalar")
pandas>=1.1.0,<3.0
